refactor: separate common components from open bus rental
