+++ src/components/items/SearchArtistItem.tsx
